Commit features: -Tests for most archive functions in csharp -Changed uint64 to uint32 due to .dll being 32bit (can't support 64 bit types). This should be fine since srcML doesn't bit shift past 31 currently -A test interface wrapper in C++ that essentially runs all of the tests and sends an answer back to csharp. This should either be a different target (separate test dll) or use #define macros to cut it out of the release version. Right now I just #include it so that it compiles into the dll; not the best way since it has to be commented out or else be included uselessly in the dll.

+++ LibSrcMLWrapper/LibSrcMLWrapper.h
@@ -21,9 +21,9 @@ namespace LibSrcMLWrapper {
         int buffercount;
         int* buffersize;
         int tabstop;
-        unsigned long long optionSet;
-        unsigned long long optionEnable;
-        unsigned long long optionDisable;
+        unsigned long optionSet;
+        unsigned long optionEnable;
+        unsigned long  optionDisable;
         char** extandlanguage;
         char** prefixandnamespace;
         char** targetanddata;
